Bugfixes and cleanup * Fixed a NRE on startup when the Version Manifest hasn't been downloaded yet. * Fixed a NRE on trying to launch the game when it hasn't been installed yet. * Fixed issue where sometimes the download form could continue downloading if user tried to cancel. * Added web browser style consistency for forum browser. * Added FxCop code analyzers and cleaned up. * Added StyleCop code analyzers and cleaned up.

+++ Apex Launcher/Forms/Launcher.cs
@@ -1,30 +1,103 @@
-﻿using System;
+﻿// <copyright file="Launcher.cs" company="IAMVISHNU Media">
+// © Copyright by IAMVISHNU Media 2020 CC BY-NC-ND
+// </copyright>
+
+using System;
 using System.Diagnostics;
 using System.IO;
 using System.Net;
 using System.Windows.Forms;
 
-namespace Apex_Launcher {
+namespace ApexLauncher {
+    /// <summary>
+    /// Main form for program.
+    /// </summary>
     public partial class Launcher : Form {
+        private bool loaded;
 
-        private bool loaded = false;
-
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Launcher"/> class.
+        /// </summary>
         public Launcher() {
             InitializeComponent();
             if (!Program.NetworkConnected) {
                 TumblrBrowser.Hide();
                 RedditBrowser.Hide();
-                //GitHubBrowser.Hide();
+                ForumBrowser.Hide();
                 TabBox.Enabled = false;
             } else {
                 NoConnectionLabel.Hide();
             }
         }
 
+        private delegate void SGV(VersionGameFiles v);
+
+        private delegate void US(string message);
+
+        /// <summary>
+        /// Enable launching the game.
+        /// </summary>
+        public void EnableLaunch() {
+            LaunchButton.Enabled = true;
+            UpdateStatus("Ready to launch");
+        }
+
+        /// <summary>
+        /// Update the launcher status message.
+        /// </summary>
+        /// <param name="message">The message to set.</param>
+        public void UpdateStatus(string message) {
+            if (StatusLabel.InvokeRequired) {
+                US d = UpdateStatus;
+                Invoke(d, new object[] { message });
+            } else StatusLabel.Text = message;
+        }
+
+        /// <summary>
+        /// Sets the current game version text.
+        /// </summary>
+        /// <param name="v">Version to set.</param>
+        public void SetGameVersion(VersionGameFiles v) {
+            if (v is null) throw new ArgumentNullException(nameof(v));
+            if (GameVersionLabel.InvokeRequired) {
+                SGV d = SetGameVersion;
+                Invoke(d, new object[] { v });
+            } else GameVersionLabel.Text = "Build: " + v.ToString();
+        }
+
+        /// <summary>
+        /// Sets the text for the launcher version.
+        /// </summary>
+        /// <param name="versionText">The text to set.</param>
+        public void SetLauncherVersion(string versionText) {
+            LauncherVersionLabel.Text = "Launcher v" + versionText;
+        }
+
+        private void SettingsButton_Click(object sender, EventArgs e) {
+            using SettingsForm settings = new SettingsForm();
+            settings.ShowDialog();
+        }
+
+        private void Launcher_FormClosing(object sender, FormClosingEventArgs e) {
+            if (Program.Downloading) {
+                DialogResult res = MessageBox.Show(
+                    "A download is still in progress. Closing the launcher will cancel your download. Are you sure you want to quit?",
+                    "Warning",
+                    MessageBoxButtons.YesNo,
+                    MessageBoxIcon.Warning);
+                if (res == DialogResult.No) e.Cancel = true;
+            }
+        }
+
+        private void SaveMgmtButton_Click(object sender, EventArgs e) {
+            using SaveManagementForm saves = new SaveManagementForm();
+            saves.ShowDialog();
+        }
+
         private void Launcher_Shown(object sender, EventArgs e) {
             TumblrBrowser.IsWebBrowserContextMenuEnabled = false;
             RedditBrowser.IsWebBrowserContextMenuEnabled = false;
-            //GitHubBrowser.IsWebBrowserContextMenuEnabled = false;
+            ForumBrowser.IsWebBrowserContextMenuEnabled = false;
             VersionGameFiles vgf = Config.CurrentVersion;
             if (vgf != null) SetGameVersion(vgf);
             LauncherVersionLabel.Text = "Launcher v" + Program.GetLauncherVersion();
@@ -35,6 +108,7 @@ namespace Apex_Launcher {
                     Program.NetworkConnected = false;
                 }
             }
+
             EnableLaunch();
         }
 
@@ -51,7 +125,7 @@ namespace Apex_Launcher {
         }
 
         private void BrowserNavigation(object sender, WebBrowserNavigatingEventArgs e) {
-            if (loaded && !(e.Url.Equals(((WebBrowser)sender).Url)) && !(e.Url.ToString().Contains("redditmedia")) && (e.Url.ToString().Contains(((WebBrowser)sender).Url.ToString()))) {
+            if (loaded && !e.Url.Equals(((WebBrowser)sender).Url) && !e.Url.ToString().Contains("redditmedia") && e.Url.ToString().Contains(((WebBrowser)sender).Url.ToString())) {
                 Process.Start(e.Url.ToString());
                 e.Cancel = true;
             } else if (loaded) {
@@ -68,15 +142,23 @@ namespace Apex_Launcher {
                 MessageBox.Show("A download is currently in progress. Please cancel your download or wait until it finishes.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                 return;
             }
-            VersionGameFiles CurrentVersion = Config.CurrentVersion;
-            string launchpath = Path.Combine(Config.InstallPath, "Versions", CurrentVersion.ToString(), "Game.exe");
+
+            VersionGameFiles currentVersion = Config.CurrentVersion;
+            if (currentVersion is null) {
+                if (MessageBox.Show("No installed version found. Would you like to download now?", "No Version Found", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) != DialogResult.Yes) return;
+
+                Program.InstallLatestVersion();
+                return;
+            }
+
+            string launchpath = Path.Combine(Config.InstallPath, "Versions", currentVersion.ToString(), "Game.exe");
 
             if (Program.ForceUpdate) {
-                string path = Config.InstallPath + "\\Versions\\" + CurrentVersion.ToString();
+                string path = Path.Combine(Config.InstallPath, "Versions", currentVersion.ToString());
                 if (File.Exists(path + ".zip")) File.Delete(path + ".zip");
                 if (Directory.Exists(path)) Directory.Delete(path, true);
-                if (CurrentVersion.IsPatch) {
-                    string previousPath = Config.InstallPath + "\\Versions\\" + CurrentVersion.Prerequisite.ToString();
+                if (currentVersion.IsPatch) {
+                    string previousPath = Path.Combine(Config.InstallPath, "Versions", currentVersion.Prerequisite.ToString());
                     if (File.Exists(previousPath + ".zip")) File.Delete(previousPath + ".zip");
                     if (Directory.Exists(previousPath)) Directory.Delete(previousPath, true);
                     Config.CurrentVersion = VersionGameFiles.FromString("ALPHA 0.0");
@@ -86,11 +168,11 @@ namespace Apex_Launcher {
                 Program.ForceUpdate = false;
             } else if (!File.Exists(launchpath)) {
                 DialogResult res = MessageBox.Show(
-                    "Cannot find the game in your install path. It might be moved or deleted.\nCheck \n" + launchpath +
-                    "\nfor your files, or redownload them.\nWould you like to redownload?", "Game not found", MessageBoxButtons.YesNo
-                );
+                    $"Cannot find the game in your install path. It might be moved or deleted.\nCheck \n{launchpath}\nfor your files, or redownload them.\nWould you like to redownload?",
+                    "Game not found",
+                    MessageBoxButtons.YesNo);
                 if (res == DialogResult.Yes) {
-                    Program.DownloadVersion(VersionGameFiles.GetMostRecentVersion());
+                    Program.InstallLatestVersion();
                 } else return;
             }
 
@@ -98,54 +180,6 @@ namespace Apex_Launcher {
                 Process.Start(launchpath);
                 if (!Config.KeepLauncherOpen) Close();
             }
-        }
-
-        public void EnableLaunch() {
-            LaunchButton.Enabled = true;
-            UpdateStatus("Ready to launch");
-        }
-
-        public delegate void US(string message);
-        public void UpdateStatus(string message) {
-            if (StatusLabel.InvokeRequired) {
-                US d = UpdateStatus;
-                Invoke(d, new object[] { message });
-            } else StatusLabel.Text = message;
-        }
-
-        public delegate void SGV(VersionGameFiles v);
-        public void SetGameVersion(VersionGameFiles v) {
-            if (GameVersionLabel.InvokeRequired) {
-                SGV d = SetGameVersion;
-                Invoke(d, new object[] { v });
-            } else GameVersionLabel.Text = "Build: " + v.ToString();
-        }
-
-        public void SetLauncherVersion(string versionText) {
-            LauncherVersionLabel.Text = "Launcher v" + versionText;
-        }
-
-        private void SettingsButton_Click(object sender, EventArgs e) {
-            SettingsForm settings = new SettingsForm();
-            settings.ShowDialog();
-        }
-
-        private void Launcher_FormClosing(object sender, FormClosingEventArgs e) {
-            if (Program.Downloading) {
-                DialogResult res = MessageBox.Show(
-                    "A download is still in progress. Closing the launcher will cancel your download. Are you sure you want to quit?", "Warning",
-                    MessageBoxButtons.YesNo,
-                    MessageBoxIcon.Warning
-                );
-                if (res == DialogResult.No) {
-                    e.Cancel = true;
-                }
-            }
-        }
-
-        private void SaveMgmtButton_Click(object sender, EventArgs e) {
-            SaveManagementForm saves = new SaveManagementForm();
-            saves.ShowDialog();
         }
     }
 }
